Add Criando usuário admin

+++ src/main/java/dao/DAOUsuarioRepository.java
@@ -18,11 +18,11 @@ public class DAOUsuarioRepository {
 		connection = SingleConnectionBanco.getConnection();
 	}
 
-	public ModelLogin gravarUsuario(ModelLogin objeto) throws Exception {
+	public ModelLogin gravarUsuario(ModelLogin objeto, Long userLogado) throws Exception {
 
 		if (objeto.isNovo()) {// Grava um novo
 
-			String sql = "INSERT INTO model_login( login, senha, nome, email) VALUES (?, ?, ?, ?);";
+			String sql = "INSERT INTO model_login( login, senha, nome, email, usuario_id) VALUES (?, ?, ?, ?, ?);";
 
 			PreparedStatement preparedStatement = connection.prepareStatement(sql);
 
@@ -30,6 +30,7 @@ public class DAOUsuarioRepository {
 			preparedStatement.setString(2, objeto.getSenha());
 			preparedStatement.setString(3, objeto.getNome());
 			preparedStatement.setString(4, objeto.getEmail());
+			preparedStatement.setLong(5, userLogado);
 
 			preparedStatement.execute();
 			connection.commit();
@@ -47,14 +48,14 @@ public class DAOUsuarioRepository {
 			connection.commit();
 		}
 
-		return this.consultarUsuario(objeto.getLogin());
+		return this.consultarUsuario(objeto.getLogin(), userLogado);
 
 	}
 
-	public List<ModelLogin> consultaUsuarioList() throws Exception {
+	public List<ModelLogin> consultaUsuarioList(Long userLogado) throws Exception {
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 
-		String sql = "SELECT * FROM model_login";
+		String sql = "SELECT * FROM model_login where useradmin is false and usuario_id = " + userLogado;
 
 		PreparedStatement preparedStatement = connection.prepareStatement(sql);
 
@@ -77,10 +78,10 @@ public class DAOUsuarioRepository {
 
 	}
 
-	public List<ModelLogin> consultaUsuarioList(String nome) throws Exception {
+	public List<ModelLogin> consultaUsuarioList(String nome, Long userLogado) throws Exception {
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 
-		String sql = "SELECT * FROM model_login WHERE upper(nome) LIKE upper(?);";
+		String sql = "SELECT * FROM model_login WHERE upper(nome) LIKE upper(?) and useradmin is false and usuario_id = " + userLogado;
 
 		PreparedStatement preparedStatement = connection.prepareStatement(sql);
 		preparedStatement.setString(1, "%" + nome + "%");
@@ -103,12 +104,63 @@ public class DAOUsuarioRepository {
 
 		return retorno;
 	}
+	
+	public ModelLogin consultarUsuarioLogado(String login) throws Exception {
+
+		ModelLogin modelLogin = new ModelLogin();
+
+		String sql = "SELECT * FROM model_login WHERE login = ?";
+
+		PreparedStatement preparedStatement = connection.prepareStatement(sql);
+		preparedStatement.setString(1, login);
+
+		ResultSet resultado = preparedStatement.executeQuery();
+
+		while (resultado.next()) { // Retorna apenas um usuário
+
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setSenha(resultado.getString("senha"));
+			modelLogin.setUseradmin(resultado.getBoolean("useradmin"));
+
+		}
+
+		return modelLogin;
+	}
+	
 
 	public ModelLogin consultarUsuario(String login) throws Exception {
 
 		ModelLogin modelLogin = new ModelLogin();
 
-		String sql = "SELECT * FROM model_login WHERE login = ? ;";
+		String sql = "SELECT * FROM model_login WHERE login = ? and useradmin is false";
+
+		PreparedStatement preparedStatement = connection.prepareStatement(sql);
+		preparedStatement.setString(1, login);
+
+		ResultSet resultado = preparedStatement.executeQuery();
+
+		while (resultado.next()) { // Retorna apenas um usuário
+
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setSenha(resultado.getString("senha"));
+			modelLogin.setUseradmin(resultado.getBoolean("useradmin"));
+
+		}
+
+		return modelLogin;
+	}
+	
+	public ModelLogin consultarUsuario(String login, Long userLogado) throws Exception {
+
+		ModelLogin modelLogin = new ModelLogin();
+
+		String sql = "SELECT * FROM model_login WHERE login = ? and useradmin is false and usuario_id = " + userLogado;
 
 		PreparedStatement preparedStatement = connection.prepareStatement(sql);
 		preparedStatement.setString(1, login);
@@ -128,11 +180,13 @@ public class DAOUsuarioRepository {
 		return modelLogin;
 	}
 
-	public ModelLogin consultarUsuarioId(String id) throws Exception {
+	
+	
+	public ModelLogin consultarUsuarioId(String id, Long userLogado) throws Exception {
 
 		ModelLogin modelLogin = new ModelLogin();
 
-		String sql = "SELECT * FROM model_login WHERE id = ? ;";
+		String sql = "SELECT * FROM model_login WHERE id = ? and useradmin is false and usuario_id = " + userLogado;
 
 		PreparedStatement preparedStatement = connection.prepareStatement(sql);
 		preparedStatement.setLong(1, Long.parseLong(id));
@@ -166,7 +220,7 @@ public class DAOUsuarioRepository {
 	}
 
 	public void deletarUsuario(String idUser) throws Exception {
-		String sql = "DELETE FROM public.model_login WHERE id = ? ;";
+		String sql = "DELETE FROM public.model_login WHERE id = ? and useradmin is false;";
 		PreparedStatement preparedStatement = connection.prepareStatement(sql);
 		preparedStatement.setLong(1, Long.parseLong(idUser));
